+++ src/label.rb
